+++ zerver/tests/test_message_send.py
@@ -3,6 +3,7 @@ from typing import Any, Optional
 from unittest import mock
 
 import ujson
+from django.http import HttpResponse
 from django.utils.timezone import now as timezone_now
 
 from zerver.decorator import JsonableError
@@ -16,10 +17,12 @@ from zerver.lib.actions import (
 from zerver.lib.message import get_recent_private_conversations
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import reset_emails_in_zulip_realm
-from zerver.lib.timestamp import datetime_to_timestamp
+from zerver.lib.timestamp import convert_to_UTC, datetime_to_timestamp
+from zerver.lib.timezone import get_timezone
 from zerver.models import (
     MAX_MESSAGE_LENGTH,
     MAX_TOPIC_NAME_LENGTH,
+    ScheduledMessage,
     Stream,
     UserProfile,
     get_huddle_recipient,
@@ -881,3 +884,137 @@ class MessagePOSTTest(ZulipTestCase):
         # Guest user can send message to subscribed public streams
         result = self.api_post(sender, "/api/v1/messages", payload)
         self.assert_json_success(result)
+
+class ScheduledMessageTest(ZulipTestCase):
+
+    def last_scheduled_message(self) -> ScheduledMessage:
+        return ScheduledMessage.objects.all().order_by('-id')[0]
+
+    def do_schedule_message(self, msg_type: str, to: str, msg: str,
+                            defer_until: str='', tz_guess: str='',
+                            delivery_type: str='send_later',
+                            realm_str: str='zulip') -> HttpResponse:
+        self.login('hamlet')
+
+        topic_name = ''
+        if msg_type == 'stream':
+            topic_name = 'Test topic'
+
+        payload = {"type": msg_type,
+                   "to": to,
+                   "client": "test suite",
+                   "content": msg,
+                   "topic": topic_name,
+                   "realm_str": realm_str,
+                   "delivery_type": delivery_type,
+                   "tz_guess": tz_guess}
+        if defer_until:
+            payload["deliver_at"] = defer_until
+
+        result = self.client_post("/json/messages", payload)
+        return result
+
+    def test_schedule_message(self) -> None:
+        content = "Test message"
+        defer_until = timezone_now().replace(tzinfo=None) + datetime.timedelta(days=1)
+        defer_until_str = str(defer_until)
+
+        # Scheduling a message to a stream you are subscribed is successful.
+        result = self.do_schedule_message('stream', 'Verona',
+                                          content + ' 1', defer_until_str)
+        message = self.last_scheduled_message()
+        self.assert_json_success(result)
+        self.assertEqual(message.content, 'Test message 1')
+        self.assertEqual(message.topic_name(), 'Test topic')
+        self.assertEqual(message.scheduled_timestamp, convert_to_UTC(defer_until))
+        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
+        # Scheduling a message for reminders.
+        result = self.do_schedule_message('stream', 'Verona',
+                                          content + ' 2', defer_until_str,
+                                          delivery_type='remind')
+        message = self.last_scheduled_message()
+        self.assert_json_success(result)
+        self.assertEqual(message.delivery_type, ScheduledMessage.REMIND)
+
+        # Scheduling a private message is successful.
+        othello = self.example_user('othello')
+        hamlet = self.example_user('hamlet')
+        result = self.do_schedule_message('private', othello.email,
+                                          content + ' 3', defer_until_str)
+        message = self.last_scheduled_message()
+        self.assert_json_success(result)
+        self.assertEqual(message.content, 'Test message 3')
+        self.assertEqual(message.scheduled_timestamp, convert_to_UTC(defer_until))
+        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
+
+        # Setting a reminder in PM's to other users causes a error.
+        result = self.do_schedule_message('private', othello.email,
+                                          content + ' 4', defer_until_str,
+                                          delivery_type='remind')
+        self.assert_json_error(result, 'Reminders can only be set for streams.')
+
+        # Setting a reminder in PM's to ourself is successful.
+        # Required by reminders from message actions popover caret feature.
+        result = self.do_schedule_message('private', hamlet.email,
+                                          content + ' 5', defer_until_str,
+                                          delivery_type='remind')
+        message = self.last_scheduled_message()
+        self.assert_json_success(result)
+        self.assertEqual(message.content, 'Test message 5')
+        self.assertEqual(message.delivery_type, ScheduledMessage.REMIND)
+
+        # Scheduling a message while guessing timezone.
+        tz_guess = 'Asia/Kolkata'
+        result = self.do_schedule_message('stream', 'Verona', content + ' 6',
+                                          defer_until_str, tz_guess=tz_guess)
+        message = self.last_scheduled_message()
+        self.assert_json_success(result)
+        self.assertEqual(message.content, 'Test message 6')
+        local_tz = get_timezone(tz_guess)
+        utz_defer_until = local_tz.normalize(local_tz.localize(defer_until))
+        self.assertEqual(message.scheduled_timestamp,
+                         convert_to_UTC(utz_defer_until))
+        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
+
+        # Test with users timezone setting as set to some timezone rather than
+        # empty. This will help interpret timestamp in users local timezone.
+        user = self.example_user("hamlet")
+        user.timezone = 'US/Pacific'
+        user.save(update_fields=['timezone'])
+        result = self.do_schedule_message('stream', 'Verona',
+                                          content + ' 7', defer_until_str)
+        message = self.last_scheduled_message()
+        self.assert_json_success(result)
+        self.assertEqual(message.content, 'Test message 7')
+        local_tz = get_timezone(user.timezone)
+        utz_defer_until = local_tz.normalize(local_tz.localize(defer_until))
+        self.assertEqual(message.scheduled_timestamp,
+                         convert_to_UTC(utz_defer_until))
+        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
+
+    def test_scheduling_in_past(self) -> None:
+        # Scheduling a message in past should fail.
+        content = "Test message"
+        defer_until = timezone_now()
+        defer_until_str = str(defer_until)
+
+        result = self.do_schedule_message('stream', 'Verona',
+                                          content + ' 1', defer_until_str)
+        self.assert_json_error(result, 'Time must be in the future.')
+
+    def test_invalid_timestamp(self) -> None:
+        # Scheduling a message from which timestamp couldn't be parsed
+        # successfully should fail.
+        content = "Test message"
+        defer_until = 'Missed the timestamp'
+
+        result = self.do_schedule_message('stream', 'Verona',
+                                          content + ' 1', defer_until)
+        self.assert_json_error(result, 'Invalid time format')
+
+    def test_missing_deliver_at(self) -> None:
+        content = "Test message"
+
+        result = self.do_schedule_message('stream', 'Verona',
+                                          content + ' 1')
+        self.assert_json_error(result, 'Missing deliver_at in a request for delayed message delivery')
