+++ zerver/tests/test_messages.py
@@ -70,8 +70,6 @@ from zerver.lib.test_helpers import (
     most_recent_usermessage,
     queries_captured,
 )
-from zerver.lib.timestamp import convert_to_UTC
-from zerver.lib.timezone import get_timezone
 from zerver.lib.topic import DB_TOPIC_NAME, TOPIC_LINKS
 from zerver.lib.types import DisplayRecipientT, UserDisplayRecipient
 from zerver.lib.upload import create_attachment
@@ -85,7 +83,6 @@ from zerver.models import (
     RealmDomain,
     RealmFilter,
     Recipient,
-    ScheduledMessage,
     Stream,
     Subscription,
     UserMessage,
@@ -1348,140 +1345,6 @@ class SewMessageAndReactionTest(ZulipTestCase):
                              'simple_smile')
             self.assertTrue(data['id'])
             self.assertTrue(data['content'])
-
-class ScheduledMessageTest(ZulipTestCase):
-
-    def last_scheduled_message(self) -> ScheduledMessage:
-        return ScheduledMessage.objects.all().order_by('-id')[0]
-
-    def do_schedule_message(self, msg_type: str, to: str, msg: str,
-                            defer_until: str='', tz_guess: str='',
-                            delivery_type: str='send_later',
-                            realm_str: str='zulip') -> HttpResponse:
-        self.login('hamlet')
-
-        topic_name = ''
-        if msg_type == 'stream':
-            topic_name = 'Test topic'
-
-        payload = {"type": msg_type,
-                   "to": to,
-                   "client": "test suite",
-                   "content": msg,
-                   "topic": topic_name,
-                   "realm_str": realm_str,
-                   "delivery_type": delivery_type,
-                   "tz_guess": tz_guess}
-        if defer_until:
-            payload["deliver_at"] = defer_until
-
-        result = self.client_post("/json/messages", payload)
-        return result
-
-    def test_schedule_message(self) -> None:
-        content = "Test message"
-        defer_until = timezone_now().replace(tzinfo=None) + datetime.timedelta(days=1)
-        defer_until_str = str(defer_until)
-
-        # Scheduling a message to a stream you are subscribed is successful.
-        result = self.do_schedule_message('stream', 'Verona',
-                                          content + ' 1', defer_until_str)
-        message = self.last_scheduled_message()
-        self.assert_json_success(result)
-        self.assertEqual(message.content, 'Test message 1')
-        self.assertEqual(message.topic_name(), 'Test topic')
-        self.assertEqual(message.scheduled_timestamp, convert_to_UTC(defer_until))
-        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
-        # Scheduling a message for reminders.
-        result = self.do_schedule_message('stream', 'Verona',
-                                          content + ' 2', defer_until_str,
-                                          delivery_type='remind')
-        message = self.last_scheduled_message()
-        self.assert_json_success(result)
-        self.assertEqual(message.delivery_type, ScheduledMessage.REMIND)
-
-        # Scheduling a private message is successful.
-        othello = self.example_user('othello')
-        hamlet = self.example_user('hamlet')
-        result = self.do_schedule_message('private', othello.email,
-                                          content + ' 3', defer_until_str)
-        message = self.last_scheduled_message()
-        self.assert_json_success(result)
-        self.assertEqual(message.content, 'Test message 3')
-        self.assertEqual(message.scheduled_timestamp, convert_to_UTC(defer_until))
-        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
-
-        # Setting a reminder in PM's to other users causes a error.
-        result = self.do_schedule_message('private', othello.email,
-                                          content + ' 4', defer_until_str,
-                                          delivery_type='remind')
-        self.assert_json_error(result, 'Reminders can only be set for streams.')
-
-        # Setting a reminder in PM's to ourself is successful.
-        # Required by reminders from message actions popover caret feature.
-        result = self.do_schedule_message('private', hamlet.email,
-                                          content + ' 5', defer_until_str,
-                                          delivery_type='remind')
-        message = self.last_scheduled_message()
-        self.assert_json_success(result)
-        self.assertEqual(message.content, 'Test message 5')
-        self.assertEqual(message.delivery_type, ScheduledMessage.REMIND)
-
-        # Scheduling a message while guessing timezone.
-        tz_guess = 'Asia/Kolkata'
-        result = self.do_schedule_message('stream', 'Verona', content + ' 6',
-                                          defer_until_str, tz_guess=tz_guess)
-        message = self.last_scheduled_message()
-        self.assert_json_success(result)
-        self.assertEqual(message.content, 'Test message 6')
-        local_tz = get_timezone(tz_guess)
-        utz_defer_until = local_tz.normalize(local_tz.localize(defer_until))
-        self.assertEqual(message.scheduled_timestamp,
-                         convert_to_UTC(utz_defer_until))
-        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
-
-        # Test with users timezone setting as set to some timezone rather than
-        # empty. This will help interpret timestamp in users local timezone.
-        user = self.example_user("hamlet")
-        user.timezone = 'US/Pacific'
-        user.save(update_fields=['timezone'])
-        result = self.do_schedule_message('stream', 'Verona',
-                                          content + ' 7', defer_until_str)
-        message = self.last_scheduled_message()
-        self.assert_json_success(result)
-        self.assertEqual(message.content, 'Test message 7')
-        local_tz = get_timezone(user.timezone)
-        utz_defer_until = local_tz.normalize(local_tz.localize(defer_until))
-        self.assertEqual(message.scheduled_timestamp,
-                         convert_to_UTC(utz_defer_until))
-        self.assertEqual(message.delivery_type, ScheduledMessage.SEND_LATER)
-
-    def test_scheduling_in_past(self) -> None:
-        # Scheduling a message in past should fail.
-        content = "Test message"
-        defer_until = timezone_now()
-        defer_until_str = str(defer_until)
-
-        result = self.do_schedule_message('stream', 'Verona',
-                                          content + ' 1', defer_until_str)
-        self.assert_json_error(result, 'Time must be in the future.')
-
-    def test_invalid_timestamp(self) -> None:
-        # Scheduling a message from which timestamp couldn't be parsed
-        # successfully should fail.
-        content = "Test message"
-        defer_until = 'Missed the timestamp'
-
-        result = self.do_schedule_message('stream', 'Verona',
-                                          content + ' 1', defer_until)
-        self.assert_json_error(result, 'Invalid time format')
-
-    def test_missing_deliver_at(self) -> None:
-        content = "Test message"
-
-        result = self.do_schedule_message('stream', 'Verona',
-                                          content + ' 1')
-        self.assert_json_error(result, 'Missing deliver_at in a request for delayed message delivery')
 
 class MessageAccessTests(ZulipTestCase):
     def test_update_invalid_flags(self) -> None:
